% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts.R
\docType{data}
\name{facts}
\alias{facts}
\title{Chuck Norris facts}
\format{A data frame with 574 rows and 1 column.}
\source{
http://www.icndb.com/
}
\usage{
data(facts, package = "roundhouse")
}
\description{
A data frame of various Chuck Norris facts. Warning, some facts contain
explicit language!
}
\examples{
head(roundhouse::facts)

# Or use
data(facts, package = "roundhouse")
head(facts)
}
\keyword{datasets}
