% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_fact.R
\name{fetch_categories}
\alias{fetch_categories}
\title{Fetch available categories}
\usage{
fetch_categories()
}
\description{
Fetch all available Chuck Norris fact categories.
}
\examples{
\dontrun{
# Fetch all available fact categories
fetch_categories()
}
}
