% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_new_fact.R
\name{suggest_new_fact}
\alias{suggest_new_fact}
\title{Suggest new Chuck Norris facts}
\usage{
suggest_new_fact(fact, user, password)
}
\arguments{
\item{fact}{Character string specifying a new Chuck Norris fact.}

\item{user}{Character string specifying a valid GitHub username.}

\item{password}{Character string specifying a valid password associated with
the above GitHub username.}
}
\description{
Have a new Chuck Norris fact you want added to the database? Use this
function to open a new issue on roundhouse's GitHub page. If it's good, I'll
be sure to include it in the next release!
}
\examples{
\dontrun{
request_new_fact}
}
