% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundhouse.R
\name{roundhouse}
\alias{roundhouse}
\title{Roundhouse}
\usage{
roundhouse(width = 15, color = "white", ...)
}
\arguments{
\item{width}{Integer specifying the number of characters per line.}

\item{color}{Character string specifying the color of the text.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[magick]{image_annotate}}.}
}
\description{
Generate a random Chuck Norris fact with a roundhouse kick.
}
\examples{
roundhouse(width = 40, size = 15)
}
